Rename of aux package to auxiliary. Aux path is reserved on windows.
